+++ src/main/java/com/thehyundai/thepet/heendycar/HeendyCarReservationVO.java
@@ -1,14 +1,12 @@
 package com.thehyundai.thepet.heendycar;
 
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
+import lombok.*;
 
 import java.time.LocalDateTime;
 
 @Getter
 @Setter
+@Builder
 @AllArgsConstructor
 @NoArgsConstructor
 public class HeendyCarReservationVO {
